refactor: extract POM cache helpers
